Percussion v0.7.0 Percussion.Command behaviour

Specification for command modules.

Types

t()

t() :: module()

Callbacks

aliases()

aliases() :: [String.t()]

Returns the list of names that this command should match on.

The first element in the list is treated as the command's primary name.

call(request)

call(request :: Percussion.Request.t()) :: Percussion.Request.t()

Executes the command on the request.

help()

help() :: String.t()

Returns the help message for this command.

usage()

usage() :: [String.t()]

Returns the usage messages for this command.