Percussion v0.7.0 Percussion.Converters

Utility functions for converting Discord mentions or strings into useful formats.

Functions

animated_emoji_to_id(text)

animated_emoji_to_id(String.t()) :: {:ok, Nostrum.Snowflake.t()} | :error

Converts a formatted Discord animated emoji to a Nostrum.Snowflake.t/0.

Examples

iex> Percussion.Converters.animated_emoji_to_id("<a:thonk:123456789>")
{:ok, 123456789}

iex> Percussion.Converters.animated_emoji_to_id("<:thonk:123456789>")
:error

iex> Percussion.Converters.animated_emoji_to_id("<@123456789>")
:error

iex> Percussion.Converters.animated_emoji_to_id("123456789")
:error

channel_to_id(text)

channel_to_id(String.t()) :: {:ok, Nostrum.Snowflake.t()} | :error

Converts a formatted Discord channel mention to a Nostrum.Snowflake.t/0.

Examples

iex> Percussion.Converters.channel_to_id("<#123456789>")
{:ok, 123456789}

iex> Percussion.Converters.channel_to_id("<@123456789>")
:error

iex> Percussion.Converters.channel_to_id("123456789")
:error

emoji_to_id(text)

Converts any formatted Discord emoji to a Nostrum.Snowflake.t/0.

Examples

iex> Percussion.Converters.emoji_to_id("<a:thonk:123456789>")
{:animated, 123456789}

iex> Percussion.Converters.emoji_to_id("<:thonk:123456789>")
{:static, 123456789}

iex> Percussion.Converters.emoji_to_id("<@123456789>")
:error

iex> Percussion.Converters.emoji_to_id("123456789")
:error

role_mention_to_id(text)

role_mention_to_id(String.t()) :: {:ok, Nostrum.Snowflake.t()} | :error

Converts a formatted Discord role mention to a Nostrum.Snowflake.t/0.

Examples

iex> Percussion.Converters.role_mention_to_id("<@&123456789>")
{:ok, 123456789}

iex> Percussion.Converters.role_mention_to_id("<@123456789>")
:error

iex> Percussion.Converters.role_mention_to_id("123456789")
:error

static_emoji_to_id(text)

static_emoji_to_id(String.t()) :: {:ok, Nostrum.Snowflake.t()} | :error

Converts a formatted Discord static emoji to a Nostrum.Snowflake.t/0.

Examples

iex> Percussion.Converters.static_emoji_to_id("<:thonk:123456789>")
{:ok, 123456789}

iex> Percussion.Converters.static_emoji_to_id("<@123456789>")
:error

iex> Percussion.Converters.static_emoji_to_id("123456789")
:error

user_mention_to_id(text)

user_mention_to_id(String.t()) :: {:ok, Nostrum.Snowflake.t()} | :error

Converts a formatted Discord user mention to a Nostrum.Snowflake.t/0.

Examples

iex> Percussion.Converters.user_mention_to_id("<@123456789>")
{:ok, 123456789}

iex> Percussion.Converters.user_mention_to_id("<@!123456789>")
{:ok, 123456789}

iex> Percussion.Converters.user_mention_to_id("123456789")
:error