Percussion v0.7.0 Percussion.Decorators

Built-in command decorators.

Functions

direct_message_only()

direct_message_only() :: Percussion.Request.step()

Requires the command to be called in a direct message.

guild_only()

guild_only() :: Percussion.Request.step()

Requires the command to be called in a guild.

log_request(level)

log_request(Logger.level()) :: Percussion.Request.step()

Logs command requests, and how long they took to complete.

measure_response_time()

measure_response_time() :: Percussion.Request.step()

Measures and stores how long a request took.

The value is stored in the request's assigns as response_time.

whitelist_guilds(whitelist)

whitelist_guilds([Nostrum.Snowflake.t()]) :: Percussion.Request.step()

Prevents the command from being called in non-whitelisted guilds.

Note that it can still be used in DMs.

whitelist_users(whitelist)

whitelist_users([Nostrum.Snowflake.t()]) :: Percussion.Request.step()

Prevents the command from being called by non-whitelisted users.

whitelisted_guilds_only(whitelist)

whitelisted_guilds_only([Nostrum.Snowflake.t()]) :: Percussion.Request.step()

Combines guild_only/1 and whitelist_guilds/1.