Percussion v0.7.0 Percussion.Request

A bot command request.

Types

after_send()

after_send() :: [step()]

Callbacks to be called right after the response is sent.

arguments()

arguments() :: [String.t()]

The list of arguments that were passed into the command.

assigns()

assigns() :: %{required(atom()) => any()}

Additional data shared between steps.

halt()

halt() :: boolean()

Whether to stop propagating this request.

into()

into() :: t() | String.t()

A value able to be converted into a request or response.

invoked_with()

invoked_with() :: String.t()

The command name that triggered this request.

message()

message() :: Nostrum.Struct.Message.t()

The message which triggered this request.

response()

response() :: String.t() | nil

The response to send to the user.

step()

step() :: (t() -> into())

A single step in the pipeline.

t()

t() :: %Percussion.Request{
  after_send: after_send(),
  arguments: arguments(),
  assigns: assigns(),
  halt: halt(),
  invoked_with: invoked_with(),
  message: message(),
  response: response()
}

Functions

and_then(request, fun)

and_then(t(), step()) :: t()

Maps request by applying a function if it's not halted, otherwise leave it untouched.

assign(request, assigns)

assign(t(), Keyword.t()) :: t()

Assigns multiple values to the request.

halt(request)

halt(t()) :: t()

Halts the pipeline, preventing downstream pipes from being executed.

halt(request, response)

halt(t(), String.t()) :: t()

Halts the pipeline with the given response, preventing downstream pipes from being executed.

Equivalent to calling halt/1 and reply/2.

map(request, fun)

map(t(), step()) :: t()

Maps request by applying fun on it.

If the value returned by fun is a string, the request is halted with the returned string as a response.

pipe(request, pipeline)

pipe(t(), [step()]) :: t()

Maps request with each element in pipeline in order.

See and_then/2. This function terminates when pipeline is exhausted, or if any of its elements halts the request.

register_after_send(request, callback)

register_after_send(t(), step()) :: t()

Registers a callback to be invoked after the response is sent.

Callbacks are invoked regardless of the request being halted, and are executed in first-in, last-out order.

reply(request, response)

reply(t(), String.t()) :: t()

Sets the response for the request.

send_response(request, callback)

send_response(t(), step()) :: t()

Sends a response to the client using callback.