Percussion v0.7.0 Percussion.Router

Command routing specification.

Types

routes()

routes() :: %{required(String.t()) => Command.t()}

Routing table.

Functions

dispatch(routes, request)

dispatch(routes(), Percussion.Request.t()) :: Percussion.Request.t()

Executes the command that matches on request.

help(routes, name)

help(routes(), String.t()) :: {:ok, String.t()} | :error

Returns the help message for the command that matches on name.

new(modules)

new([Command.t()]) :: routes()

Builds a routing table for the given modules.

put(routes, route)

put(routes(), Command.t()) :: routes()

Inserts route into the routing table.

resolve(routes, name)

resolve(routes(), String.t()) :: {:ok, Command.t()} | :error

Performs a lookup by name on router.

usage(routes, name)

usage(routes(), String.t()) :: {:ok, [String.t()]} | :error

Returns the usage messages for the command that matches on name.