Percussion v0.7.0 Percussion.Utils

Utility functions that don't warrant their own modules, but are likely useful for most bots.

Functions

create_message(request)

create_message(Percussion.Request.t()) ::
  {:ok, Nostrum.Struct.Message.t()} | Nostrum.Api.error()

Replies a request to the respective channel.

split(text)

split(String.t()) :: [String.t()]

Tries splitting the given text with OptionParser.split/1, and falls back to String.split/1 if it fails.

Examples

iex> Percussion.Utils.split(~s[choose "lorem ipsum" "hello world"])
["choose", "lorem ipsum", "hello world"]

iex> Percussion.Utils.split(~s[I'm a bot])
["I'm", "a", "bot"]

to_request(message, contents)

to_request(Nostrum.Struct.Message.t(), String.t()) :: Percussion.Request.t()

Creates a Percussion.Request.t/0 from a Nostrum.Struct.Message.t/0.

Argument contents must be the unprefixed contents of the message that triggered this command. Note that enforcing that message and contents matches is up to the caller.